Refactor package to davidmigloz
