+++ lib/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="com.davidmiguel.numberkeyboard"/>
+<manifest package="com.davidmigloz.numberkeyboard"/>
