+++ lib/src/main/kotlin/com/davidmigloz/numberkeyboard/listener/NumberKeyboardListener.kt
@@ -1,0 +1,7 @@
+package com.davidmigloz.numberkeyboard.listener
+
+import com.davidmigloz.numberkeyboard.data.NumberKeyboardData
+
+interface NumberKeyboardListener {
+    fun onUpdated(data: NumberKeyboardData)
+}
