+++ sample/src/androidMain/kotlin/com/davidmigloz/numberkeyboard/sample/BiometricScreen.kt
@@ -1,18 +1,17 @@
-package com.davidmiguel.numberkeyboard.sample
+package com.davidmigloz.numberkeyboard.sample
 
+import android.content.Context
+import android.widget.Toast
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.consumeWindowInsets
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeGesturesPadding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.rounded.Backspace
+import androidx.compose.material.icons.rounded.Fingerprint
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -21,24 +20,21 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
-import com.davidmiguel.numberkeyboard.NumberKeyboard
-import com.davidmiguel.numberkeyboard.NumberKeyboardAuxButton
-import com.davidmiguel.numberkeyboard.NumberKeyboardButton
-import com.davidmiguel.numberkeyboard.data.NumberKeyboardData
-import com.davidmiguel.numberkeyboard.listener.NumberKeyboardListener
-import java.text.DecimalFormat
-import java.text.NumberFormat
+import com.davidmigloz.numberkeyboard.NumberKeyboard
+import com.davidmigloz.numberkeyboard.NumberKeyboardAuxButton
+import com.davidmigloz.numberkeyboard.NumberKeyboardButton
+import com.davidmigloz.numberkeyboard.data.NumberKeyboardData
+import com.davidmigloz.numberkeyboard.listener.NumberKeyboardListener
 
 @Composable
-fun CustomScreen(innerPadding: PaddingValues) {
-    val currencySymbol = "€"
+fun BiometricScreen(innerPadding: PaddingValues, context: Context = LocalContext.current) {
     Column(
         Modifier
             .padding(innerPadding)
             .consumeWindowInsets(innerPadding)
             .safeGesturesPadding()
-            .verticalScroll(rememberScrollState())
     ) {
         Text(
             text = "Number Keyboard",
@@ -46,23 +42,22 @@ fun CustomScreen(innerPadding: PaddingValues) {
         )
 
         Text(
-            text = "Custom",
+            text = "Fingerprint",
             style = MaterialTheme.typography.titleLarge,
         )
 
         Spacer(modifier = Modifier.height(8.dp))
 
         Text(
-            text = "maxAllowedAmount = 8_888.888\n" +
-                    "maxAllowedDecimals = 3\n" +
-                    "currencySymbol = $currencySymbol\n" +
-                    "isInverted = true",
+            text = "maxAllowedAmount = 9_999.00\n" +
+                    "maxAllowedDecimals = 0\n" +
+                    "roundUpToMax = false",
             style = MaterialTheme.typography.titleMedium,
         )
 
         Spacer(modifier = Modifier.height(16.dp))
 
-        var text by remember { mutableStateOf("$currencySymbol 0") }
+        var text by remember { mutableStateOf("0") }
 
         Text(
             text = text,
@@ -73,45 +68,39 @@ fun CustomScreen(innerPadding: PaddingValues) {
 
         val buttonModifier = Modifier
             .weight(1F)
-            .aspectRatio(1F)
+            .height(48.dp)
         val buttonTextStyle = MaterialTheme.typography.titleMedium
         NumberKeyboard(
-            maxAllowedAmount = 8_888.888,
-            maxAllowedDecimals = 3,
-            currencySymbol = currencySymbol,
-            isInverted = true,
+            maxAllowedAmount = 9_999.00,
+            maxAllowedDecimals = 0,
+            roundUpToMax = false,
             button = { number, clickedListener ->
                 NumberKeyboardButton(
                     modifier = buttonModifier,
                     textStyle = buttonTextStyle,
-                    shape = CircleShape,
                     number = number,
                     listener = clickedListener
                 )
             },
-            leftAuxButton = { clickedListener ->
+            leftAuxButton = { _ ->
                 NumberKeyboardAuxButton(
                     modifier = buttonModifier,
                     textStyle = buttonTextStyle,
-                    shape = CircleShape,
-                    value = (NumberFormat.getNumberInstance() as DecimalFormat).decimalFormatSymbols.decimalSeparator.toString(),
-                    clicked = { clickedListener.onLeftAuxButtonClicked() }
+                    imageVector = Icons.Rounded.Fingerprint,
+                    clicked = { Toast.makeText(context, "Biometrics Triggered", Toast.LENGTH_SHORT).show() }
                 )
             },
             rightAuxButton = { clickedListener ->
                 NumberKeyboardAuxButton(
                     modifier = buttonModifier,
                     textStyle = buttonTextStyle,
-                    shape = CircleShape,
                     imageVector = Icons.AutoMirrored.Rounded.Backspace,
                     clicked = { clickedListener.onRightAuxButtonClicked() }
                 )
             },
-            decimalSeparator = (NumberFormat.getNumberInstance() as DecimalFormat).decimalFormatSymbols.decimalSeparator,
-            groupingSeparator = (NumberFormat.getNumberInstance() as DecimalFormat).decimalFormatSymbols.groupingSeparator,
             listener = object : NumberKeyboardListener {
                 override fun onUpdated(data: NumberKeyboardData) {
-                    text = data.currency
+                    text = data.int.toString()
                 }
             }
         )
