+++ sample/src/androidMain/kotlin/com/davidmigloz/numberkeyboard/sample/CustomScreen.kt
@@ -1,12 +1,16 @@
-package com.davidmiguel.numberkeyboard.sample
+package com.davidmigloz.numberkeyboard.sample
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.consumeWindowInsets
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeGesturesPadding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.rounded.Backspace
 import androidx.compose.material3.MaterialTheme
@@ -18,22 +22,23 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.davidmiguel.numberkeyboard.NumberKeyboard
-import com.davidmiguel.numberkeyboard.NumberKeyboardAuxButton
-import com.davidmiguel.numberkeyboard.NumberKeyboardButton
-import com.davidmiguel.numberkeyboard.data.NumberKeyboardData
-import com.davidmiguel.numberkeyboard.listener.NumberKeyboardListener
+import com.davidmigloz.numberkeyboard.NumberKeyboard
+import com.davidmigloz.numberkeyboard.NumberKeyboardAuxButton
+import com.davidmigloz.numberkeyboard.NumberKeyboardButton
+import com.davidmigloz.numberkeyboard.data.NumberKeyboardData
+import com.davidmigloz.numberkeyboard.listener.NumberKeyboardListener
 import java.text.DecimalFormat
 import java.text.NumberFormat
 
 @Composable
-fun DecimalScreen(innerPadding: PaddingValues) {
-    val currencySymbol = "$"
+fun CustomScreen(innerPadding: PaddingValues) {
+    val currencySymbol = "€"
     Column(
         Modifier
             .padding(innerPadding)
             .consumeWindowInsets(innerPadding)
             .safeGesturesPadding()
+            .verticalScroll(rememberScrollState())
     ) {
         Text(
             text = "Number Keyboard",
@@ -41,8 +46,18 @@ fun DecimalScreen(innerPadding: PaddingValues) {
         )
 
         Text(
-            text = "Decimal",
+            text = "Custom",
             style = MaterialTheme.typography.titleLarge,
+        )
+
+        Spacer(modifier = Modifier.height(8.dp))
+
+        Text(
+            text = "maxAllowedAmount = 8_888.888\n" +
+                    "maxAllowedDecimals = 3\n" +
+                    "currencySymbol = $currencySymbol\n" +
+                    "isInverted = true",
+            style = MaterialTheme.typography.titleMedium,
         )
 
         Spacer(modifier = Modifier.height(16.dp))
@@ -58,13 +73,18 @@ fun DecimalScreen(innerPadding: PaddingValues) {
 
         val buttonModifier = Modifier
             .weight(1F)
-            .height(48.dp)
+            .aspectRatio(1F)
         val buttonTextStyle = MaterialTheme.typography.titleMedium
         NumberKeyboard(
+            maxAllowedAmount = 8_888.888,
+            maxAllowedDecimals = 3,
+            currencySymbol = currencySymbol,
+            isInverted = true,
             button = { number, clickedListener ->
                 NumberKeyboardButton(
                     modifier = buttonModifier,
                     textStyle = buttonTextStyle,
+                    shape = CircleShape,
                     number = number,
                     listener = clickedListener
                 )
@@ -73,6 +93,7 @@ fun DecimalScreen(innerPadding: PaddingValues) {
                 NumberKeyboardAuxButton(
                     modifier = buttonModifier,
                     textStyle = buttonTextStyle,
+                    shape = CircleShape,
                     value = (NumberFormat.getNumberInstance() as DecimalFormat).decimalFormatSymbols.decimalSeparator.toString(),
                     clicked = { clickedListener.onLeftAuxButtonClicked() }
                 )
@@ -81,10 +102,13 @@ fun DecimalScreen(innerPadding: PaddingValues) {
                 NumberKeyboardAuxButton(
                     modifier = buttonModifier,
                     textStyle = buttonTextStyle,
+                    shape = CircleShape,
                     imageVector = Icons.AutoMirrored.Rounded.Backspace,
                     clicked = { clickedListener.onRightAuxButtonClicked() }
                 )
             },
+            decimalSeparator = (NumberFormat.getNumberInstance() as DecimalFormat).decimalFormatSymbols.decimalSeparator,
+            groupingSeparator = (NumberFormat.getNumberInstance() as DecimalFormat).decimalFormatSymbols.groupingSeparator,
             listener = object : NumberKeyboardListener {
                 override fun onUpdated(data: NumberKeyboardData) {
                     text = data.currency
